pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["nurbol0tt <fairtail764@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.19"
pydantic = {extras = ["email"], version = "^2.1.1"}
pydantic-settings = "^2.0.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
asyncpg = "^0.28.0"
alembic = "^1.11.1"
passlib = "^1.7.4"
async-fastapi-jwt-auth = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
